Update Review.java with totalSentiment method.
The method totalSentiment works by finding the sentiment value of each word; making sure to exclude any punctuation because it does not contribute to the overall value of it. Afterwards, we add letters together until the method reaches a space, where a method will be called to find the sentiment value of the overall word found. The value is stored and added into sentiment and finally returned by totalSentiment.